% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Concrete}
\alias{Concrete}
\title{Concrete Compressive Strength Data}
\format{
\code{Concrete} is a data frame with the following 
variables.
\itemize{ 
\item{limestone}{ percentage of limestone}
\item{water}{ water-cement ratio}
\item{strength}{ compressive strength (MPa) after 28 days}
}
}
\description{
These data were collected by I-Cheng Yeh to determine how the compressive
strength of concrete is related to its ingredients (cement, blast furnace
slag, fly ash, water, superplasticizer, coarse aggregate, and fine
aggregate) and age.
}
\references{
Appeared in Devore's "Probability and Statsistics for Engineers and 
the Sciences (6th ed).  The variables have been renamed.
}
