% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastR2-package.R
\docType{package}
\name{fastR2-package}
\alias{fastR2}
\alias{fastR2-package}
\alias{fastR}
\title{fastR2: Foundations and Applications of Statistics Using R (2nd Edition)}
\description{
Data sets and utilities to accompany the second edition of "Foundations and Applications of Statistics: an Introduction using R" (R Pruim, published by AMS, 2017), a text covering topics from probability and mathematical statistics at an advanced undergraduate level. R is integrated throughout, and access to all the R code in the book is provided via the snippet() function.

Data sets and utility functions to accompany \cite{Foundations and
Applications of Statistics: An Introduction Using R} by Randall Pruim.
}
\examples{

require(fastR2)
trellis.par.set(theme=col.fastR())

}
\references{
R. Pruim, \cite{Foundations and Applicaitons of Statistics: An Introduction Using R}, AMS, 2011.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rpruim/fastR2}
  \item \url{http://rpruim.github.io/fastR2/}
  \item Report bugs at \url{https://github.com/rpruim/fastR2/issues}
}

}
\author{
\strong{Maintainer}: Randall Pruim \email{rpruim@calvin.edu}


Randall Pruim

Maintainer: Randall Pruim <rpruim@calvin.edu>
}
\keyword{internal}
\keyword{package}
