% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLik2.R
\name{information}
\alias{information}
\title{Information}
\usage{
information(object, ...)
}
\arguments{
\item{object}{an object of class \code{"maxLik"}.}

\item{...}{additional arguments}
}
\description{
Extract information from a maxLik object
}

