% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLik2.R
\name{maxLik2}
\alias{maxLik2}
\title{Augmented version of maxLik}
\usage{
maxLik2(loglik, ..., env = parent.frame())
}
\arguments{
\item{loglik}{a log-likelihood function as for \code{\link{maxLik}}}

\item{...}{additional arguments passed to \code{\link{maxLik}}}

\item{env}{an environment in which to evaluate \code{loglik}.}
}
\description{
This version of \code{\link{maxLik}} stores additional information in the 
returned object enabling a plot method.
}
