% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NCAA2010}
\alias{NCAA2008}
\alias{NCAA2009}
\alias{NCAA2010}
\title{NCAA Division I Basketball Results}
\format{Seven variables describing NCAA Division I basketball games.
\itemize{ \item{date}{ date on which game was played}
\item{away}{ visiting team} \item{ascore}{ visiting team's
score} \item{home}{ home team} \item{hscore}{ home team's
score} \item{notes}{ code indicting games played at neutral sites (n
or N) or in tournaments (T)} \item{location}{ where game was played}
}}
\source{
\url{kenpom.com}
}
\description{
Results of NCAA basketball games
}
\examples{

data(NCAA2010)
# add some additional variables to the data frame
NCAA2010$dscore <- NCAA2010$hscore- NCAA2010$ascore
NCAA2010$homeTeamWon <- NCAA2010$dscore > 0
NCAA2010$numHomeTeamWon <- -1 + 2 * as.numeric(NCAA2010$homeTeamWon)
w <- which(NCAA2010$homeTeamWon)
NCAA2010$winner <- as.character(NCAA2010$away)
NCAA2010$winner[w] <- as.character(NCAA2010$home)[w]
NCAA2010$loser <- as.character(NCAA2010$home)
NCAA2010$loser[w] <- as.character(NCAA2010$away)[w]
NCAA2010$homeTeamWon <- NCAA2010$winner == NCAA2010$home
NCAA2010$numHomeTeamWon <- -1 + 2 * as.numeric(NCAA2010$homeTeamWon)

}
\keyword{datasets}

