% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLik2.R
\name{plot.maxLik2}
\alias{plot.maxLik2}
\title{plot method for augment maxLik objects}
\usage{
\method{plot}{maxLik2}(x, y, ci = "Wald", hline = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"maxLik2"}}

\item{y}{ignored}

\item{ci}{a character vector with values among
\code{"Wald"} and \code{"likelihood"} specifying the type of 
interval to display}

\item{hline}{a logical indicating whether a horizontal line should 
be added}

\item{...}{additional arguments, currently ignored.}
}
\description{
See [maxLik2()] and [maxLik::maxLik()] for how to create
the objects this method prints.
}
