\name{ratpoison}
\alias{ratpoison}
\docType{data}
\title{ Rat poison -- unfinished documentation }
\description{
Data from an experiment to see whether flavor and location of 
rat poison influence the consumption by rats.
}
\usage{data(ratpoison)}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{consumption}}{a numeric vector}
    \item{\code{flavor}}{a factor with levels \code{bread} \code{butter-vanilla} \code{plain} \code{roast beef}}
    \item{\code{location}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
  }
}
\examples{
data(ratpoison)
}
\keyword{datasets}
