% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rgolfballs}
\alias{rgolfballs}
\title{Simulated golf ball data}
\description{
A matrix of random golf ball numbers simulated using
\code{rmultinom(n=10000,size=486,prob=rep(0.25,4))}.
}
\examples{

data(rgolfballs)

}
\keyword{datasets}
