\name{spheres}
\alias{spheres}
\docType{data}
\title{ Measuring spheres }
\description{
Measurements of the diameter (in meters) and mass (in kilograms) of a set of steel ball bearings.  
}
\usage{data(spheres)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{diameter}}{diameter of bearing (m)}
    \item{\code{mass}}{mass of the bearing (kg) }
  }
}
\source{
These data were collected by Calvin College physics students under the direction of Steve Plath.
}
\references{
data(spheres)
}
\keyword{datasets}
