\name{undocumented}
\alias{tdf}
\alias{dotPlot}
\alias{panel.dotPlot}
\alias{funplot}
\alias{grid.identify.points}
\alias{ntigram}
\alias{panel.lm}
\alias{panel.lmbands}
\alias{panel.smooth}
\alias{panel.xyplotpoints}
\alias{panel.xyplotsmooth}
\alias{pdfplot}
\alias{pmfhistogram}
\alias{pmfplot}
\alias{probplot}
\alias{stemplot}
\alias{stemplot.data.frame}
\alias{stemplot.numeric}
\alias{funvec}
\alias{givenOrder}
\alias{make.rm}
\alias{multirandom}
\alias{nlmax}
\alias{nlmin}
\alias{summary.nlmax}
\alias{summary.nlmin}
\alias{print.stemplot}
\alias{show.fumbles}
\alias{show.poisson}
\alias{statTally}
\alias{surround}
\title{Undocumented functions}
\description{Documentation is still pending for this functions}
\author{Randall Pruim}

