% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ACTgpa}
\alias{ACTgpa}
\title{ACT scores and GPA}
\format{A data frame with 26 observations on the following 2 variables.
\itemize{ \item{ACT}{ACT score} \item{GPA}{GPA} }}
\description{
ACT scores and college GPA for a small sample of college students.
}
\examples{

xyplot(GPA ~ ACT, data = ACTgpa)

}
\keyword{datasets}

