% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AirPollution}
\alias{AirPollution}
\title{Air pollution measurements}
\format{
A data frame with 6 observations on the following 2 variables.
\describe{ 
  \item{pollution}{ a numeric vector}
  \item{location}{ a factor with levels \code{Hill Suburb}, 
        \code{Plains Suburb}, \code{Urban City}} }
}
\source{
David J. Saville and Graham R. Wood, \emph{Statistical methods: A
geometric primer}, Springer, 1996.
}
\description{
Air pollution measurements at three locations.
}
\examples{

data(AirPollution)
summary(lm(pollution ~ location, data = AirPollution))

}
\keyword{datasets}
