% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{batting}
\alias{batting}
\title{Major League Batting 2000-2005}
\format{A data frame with 8062 observations on the following 22 variables.
\itemize{ \item{player}{ unique identifier for each player}
\item{year}{ year} \item{stint}{ for players who were traded
mid-season, indicates which portion of the season the data cover}
\item{team}{ three-letter code for team} \item{league}{ a
factor with levels \code{AA} \code{AL} \code{NL}} \item{G}{ games}
\item{AB}{ at bats} \item{R}{ runs} \item{H}{ hits}
\item{H2B}{ doubles} \item{H3B}{ triples}
\item{HR}{ home runs} \item{RBI}{ runs batted in}
\item{SB}{ stolen bases} \item{CS}{ caught stealing}
\item{BB}{ bases on balls (walks)} \item{SO}{ strike outs}
\item{IBB}{ intentional base on balls} \item{HBP}{ hit by
pitch} \item{SH}{ a numeric vector} \item{SF}{ sacrifice fly}
\item{GIDP}{ grounded into double play} }}
\description{
Major League batting data for the seasons from 2000-2005.
}
\examples{

data(batting)
histogram(~HR,batting)

}
\keyword{datasets}

