% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{births78}
\alias{births78}
\title{Births by Day}
\format{A data frame with 365 observations on the following 3 variables.
\itemize{ \item{date}{ date (as a factor)}
\item{births}{ number of live births} \item{dayofyear}{ number
of days since start of year} }}
\description{
The number of live births in the United States for each day of the year
1978.
}
\examples{

data(births78)
xyplot(births ~ dayofyear, births78)

}
\keyword{datasets}
