% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{buckthorn}
\alias{buckthorn}
\title{Buckthorn}
\format{A data frame with 165 observations on the following 3 variables.
\itemize{ \item{shoots}{ number of new shoots coming from stump}
\item{conc}{ concentration of glyphosate applied}
\item{dead}{ weather the stump was considered dead} }}
\source{
David Dornbos, Calvin College
}
\description{
Data from an experiment to determine the efficacy of various methods of
eradicating buckthorn, an invasive woody shrub.  Buckthorn plants were
chopped down and the stumps treated with various concentrations of
glyphosate.  The next season, researchers returned to see whether the plant
had regrown.
}
\examples{
data(buckthorn)
}
\keyword{datasets}

