% Generated by roxygen2 (4.0.1): do not edit by hand
\name{col.fastR}
\alias{col.fastR}
\title{Lattice Theme}
\usage{
col.fastR(bw = FALSE, lty = 1:7)
}
\arguments{
\item{bw}{whether color scheme should be "black and white"}

\item{lty}{vector of line type codes}
}
\value{
Returns a list that can be supplied as the \code{theme} to
\code{\link{trellis.par.set}()}.
}
\description{
A theme for use with lattice graphics.
}
\note{
This theme was used in the production of the book \cite{Foundations
and Applications of Statistics}
}
\examples{
trellis.par.set(theme=col.fastR(bw=TRUE))
show.settings()
trellis.par.set(theme=col.fastR())
show.settings()
}
\author{
Randall Pruim
}
\seealso{
\code{\link{trellis.par.set}}, \code{\link{show.settings}}
}
\keyword{graphics}

