% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.perc.R
\name{col.perc}
\alias{col.perc}
\alias{row.perc}
\alias{row.perc}
\title{Row and Column Percentages}
\usage{
col.perc(x)

row.perc(x)
}
\arguments{
\item{x}{matrix-like structure}
}
\description{
Convenience wrappers around \code{\link{apply}()} to compute row and column
percentages of matrix-like structures, including output of
\code{\link{xtabs}}.
}
\examples{

row.perc(tally(~ airline + result, data = AirlineArrival))
col.perc(tally(~ airline + result, data = AirlineArrival))

}
\author{
Randall Pruim
}
\keyword{array}
