% Generated by roxygen2 (4.0.0): do not edit by hand
\name{col.perc}
\alias{col.perc}
\alias{row.perc}
\title{Row and Column Percentages}
\usage{
col.perc(x)

row.perc(x)
}
\arguments{
  \item{x}{matrix-like structure}
}
\description{
Convenience wrappers around \code{\link{apply}()} to compute row and column
percentages of matrix-like structures, including output of
\code{\link{xtabs}}.
}
\examples{
row.perc(xtabs(~Airline+Result,airlineArrival))
}
\author{
Randall Pruim
}
\keyword{array}

