% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cuckoo}
\alias{Cuckoo}
\title{Cuckoo eggs in other birds' nests}
\format{
A data frame with 120 observations on the following 2 variables.
\itemize{ \item{length}{ length of egg (mm)}
\item{species}{ a factor with levels \code{hedge sparrow}
\code{meadow pipet} \code{pied wagtail} \code{robin} \code{tree pipet}
\code{wren}} }
}
\source{
L.H.C. Tippett, \emph{The Methods of Statistics}, 4th Edition, John
Wiley and Sons, Inc., 1952, p. 176.
}
\description{
Cuckoos are knows to lay their eggs in the nests of other (host) birds.  The
eggs are then adopted and hatched by the host birds. These data were
originally collected by O. M. Latter in 1902 to see how the size of a cuckoo
egg is related to the species of the host bird.
}
\examples{

data(Cuckoo)
gf_boxplot(length ~ species, data = Cuckoo)

}
\references{
These data are also available from DASL, the data and story
library 
(\url{https://dasl.datadescription.com/}).
}
\keyword{datasets}
