% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{deathPenalty}
\alias{deathPen}
\alias{deathPenalty}
\title{Death Penalty and Race}
\format{A data frame with 326 observations.  The factors are coded more
succinctly in \code{deathPen}, but otherwise the data are the same.
\itemize{ \item{DeathPenalty}{ a factor with levels \code{Yes}
\code{No}} \item{Penalty}{ a factor with levels \code{Death}
\code{Not}} \item{Victim}{ a factor with levels \code{Black}
\code{White} (or \code{Bl} \code{Wh})} \item{Defendant}{ a factor
with levels \code{Black} \code{White} (or \code{Bl} \code{Wh})} }}
\source{
Radelet, M. (1981). Racial characteristics and imposition of the
death penalty. \emph{American Sociological Review}, 46:918--927.
}
\description{
A famous example of Simpson's paradox.
}
\examples{
xtabs(~Defendant+Penalty,deathPenalty)
xtabs(~Defendant+Victim+Penalty,deathPenalty)
}
\keyword{datasets}

