% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastR-package.R
\docType{package}
\name{fastR-package}
\alias{fastR}
\alias{fastR-package}
\title{Foundations and Applications of Statistics}
\description{
Data sets and utility functions to accompany \cite{Foundations and
Applications of Statistics: An Introduction Using R} by Randall Pruim.
}
\examples{
require(fastR)
trellis.par.set(theme=col.fastR())
}
\author{
Randall Pruim

Maintainer: Randall Pruim <rpruim@calvin.edu>
}
\references{
R. Pruim, \cite{Foundations and Applicaitons of Statistics: An Introduction Using R}, AMS, 2011.
}
\keyword{package}

