% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{golfballs}
\alias{golfballs}
\title{Golf ball numbers}
\format{The format is: num [1:4] 137 138 107 104}
\source{
Data collected by Allan Rossman in Carlisle, PA.
}
\description{
Allan Rossman used to live on a golf course in a spot where dozens of balls
would come into his yard every week.  He collected the balls and eventually
tallied up the numbers on the first 5000 golf balls he collected. Of these
486 bore the number 1, 2, 3, or 4.  The remaining 14 golf balls were omitted
from the data.
}
\examples{
data(golfballs)
golfballs/sum(golfballs)
chisq.test(golfballs, p=rep(.25,4))
}
\keyword{datasets}

