% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GPA}
\alias{GPA}
\title{GPA, ACT, and SAT scores}
\format{A data frame with 271 observations on the following 4 variables.
\itemize{ \item{satm}{ SAT mathematics score}
\item{satv}{ SAT verbal score} \item{act}{ ACT score}
\item{GPA}{ college grade point average} }}
\description{
GPA, ACT, and SAT scores for a sample of students.
}
\examples{

data(GPA)
splom(GPA)

}
\keyword{datasets}

