% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Jordan8687}
\alias{Jordan8687}
\title{Michael Jordan personal scoring}
\format{A data frame with 82 observations on the following 2 variables.
\itemize{ \item{Game}{a numeric vector} \item{Points}{a
numeric vector} }}
\description{
The number of points scored by Michael Jordan in each game of the 1986-87
regular season.
}
\examples{
data(Jordan8687)
xqqmath(~Points,data=Jordan8687)
}
\keyword{datasets}

