% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Jordan8687}
\alias{Jordan8687}
\title{Michael Jordan personal scoring}
\format{A data frame with 82 observations on the following 2 variables.
\itemize{ 
  \item{game}{a numeric vector} 
  \item{points}{a numeric vector} }}
\description{
The number of points scored by Michael Jordan in each game of the 1986-87
regular season.
}
\examples{

data(Jordan8687)
qqmath(~ points, data = Jordan8687)

}
\keyword{datasets}
