% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LittleSurvey}
\alias{LittleSurvey}
\title{Results from a little survey}
\format{
A data frame with 279 observations on the following 20 variables.
\describe{ 
\item{number}{ a number between 1 and 30}
\item{colorVer}{ which version of the 'favorite color' question was
on the survey. A factor with levels \code{v1} \code{v2}}
\item{color}{ favorite color if among predefined choices.  A factor
with levels \code{} \code{black} \code{green} \code{other} \code{purple}
\code{red}} 
\item{otherColor}{ favorite color if not among choices
above.} 
\item{animalVer}{ which version of the 'favorite color'
question was on the survey. A factor with levels \code{v1} \code{v2}}
\item{animal}{ favorite animal if among predefined choices.  A factor
with levels \code{} \code{elephant} \code{giraffe} \code{lion}
\code{other}.} 
\item{otherAnimal}{ favorite animal if not among the
predefined choices.} 
\item{pulseVer}{ which version of the 'pulse'
question was on the survey} 
\item{pulse}{ self-reported pulse}
\item{TVver}{ which of three versions of the TV question was on the
survey} 
\item{tvBox}{ a factor with levels \code{<1} \code{>4}
\code{>8} \code{1-2} \code{2-4} \code{4-8} \code{none} \code{other}}
\item{tvHours}{ a numeric vector} 
\item{surprisever}{ which of
two versions of the 'surprise' question was on the survey}
\item{surprise}{ a factor with levels \code{no} \code{yes}}
\item{playver}{ which of two versions of the 'play' question was on
the survey} 
\item{play}{ a factor with levels \code{no} \code{yes}}
\item{diseasever}{ which of two versions of the 'play' question was
on the survey} 
\item{disease}{ a factor with levels \code{A}
\code{B}} 
\item{homeworkver}{ which of two versions of the 'homework'
question was on the survey} 
\item{homework}{ a factor with levels
\code{A} \code{B}} }
}
\description{
These data are from a little survey given to a number of students in
introductory statistics courses. Several of the items were prepared in
multiple versions and distributed randomly to the students.
}
\section{Question Wording}{


1.1. Write down any number between 1 and 30 (inclusive).

2.1. What is your favorite color? Choices: black red; green; purple; other

2.2. What is your favorite color?

3.1. What is your favorite zoo animal? Choices: giraffe; lion; elephant; other

3.2. What is your favorite zoo animal?

4.1. Measure and record your pulse.

5.1. How much time have you spent watching TV in the last week?

5.2. How much time have you spent watching TV in the last week? 
Choises: none; under 1; hour 1-2 hours; 2-4 hours; more than 4 hours

5.3. How much time have you spent watching TV in the last week? 
Choises: under 1 hour; 1-2 hours; 2-4 hours; 4-8 hours;  more than 8 hours

6.1. Social science researchers have conducted extensive empirical studies
  and concluded that the expression "absence makes the heart grow fonder" is
  generally true. Do you find this result surprising or not surprising?
  
6.2. Social science researchers have conducted extensive empirical studies
   and concluded that the expression "out of sight out of mind" is generally
   true. Do you find this result surprising or not surprising?
   
7.1. Suppose that you have decided to see a play for which the admission charge is $20 per ticket. 
    As you prepare to purchase the ticket, you discover that you have lost a $20 bill. 
    Would you still pay $20 for a ticket to see the play?
    
7.2. Suppose that you have decided to see a play for which the admission charge is $20 per ticket. 
    As you prepare to enter the theater, you discover that you have lost your ticket. 
    Would you pay $20 to buy a new ticket to see the play?
    
8.1. suppose that the United States is preparing for the outbreak of an unusual Asian disease that is 
    expected to kill 600 people. Two alternative programs to combat the disease have been proposed. 
    Assume that the exact scientific estimates of the consequences of the programs are as follows:
    If program A is adopted, 200 people will be saved.
    If program B is adopted, there is a 1/3 probability that 600 people will be saved 
    and a 2/3 probability that nobody will be saved.
    Which of the two programs would you favor? 
    
8.2. Suppose that the United States is preparing for the outbreak of an unusual Asian disease that is 
    expected to kill 600 people. two alternative programs to combat the disease have been proposed. 
    Assume that the exact scientific estimates of the consequences of the programs are as follows:
    
    If program A is adopted, 400 people will die.
    If program B is adopted, there is a 1/3 probability that no one will die and a 2/3 probability
    that all 600 people will die.
    Which of the two programs would you favor? A or B

9.1. A national survey of college students revealed that professors at this college assign 
"significantly more homework that the nationwide average for an institution of its type."
How does this finding compare with your experience?  
    Choises: a. That sounds about right to me; b that doesn't sound right to me.
    
9.2. A national survey of college students revealed that professors at this college assign 
an amount of homework that "is fairly typical for institutions of its type."
How does this finding compare with your experience?
Choices: A that sounds about right to me; b that doesn't sound right to me.
}

\examples{
data(LittleSurvey)
tally(surprise ~ surprisever, data = LittleSurvey)
tally(disease ~ diseasever, data = LittleSurvey)

}
\keyword{datasets}
