% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mathnoise}
\alias{mathnoise}
\title{Test performance and noise}
\format{A data frame with 40 observations on the following 3 variables.
\itemize{ \item{score}{ score on a mathematics test}
\item{noise}{ a factor with levels \code{hi} \code{lo}}
\item{group}{ a factor with levels \code{control} \code{hyper}} }}
\source{
Sydney S. Zentall and Jandira H. Shaw, Effects of classroom noise on
perfor- mance and activity of second-grade hyperactive and control children,
\emph{Journal of Educational Psychology} 72 (1980), no. 6, 830.
}
\description{
In this experiment, hyperactive and control students were given a
mathematics test in either a quiet or loud testing environment.
}
\examples{
data(mathnoise)
xyplot(score~noise, data=mathnoise, group=group, type='a',
		auto.key=list(columns=2, lines=TRUE, points=FALSE))
}
\keyword{datasets}

