% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{miaa05}
\alias{miaa05}
\title{MIAA basketball 2004-2005 season}
\format{A data frame with 134 observations on the following 27 variables.
\itemize{ \item{Number}{ jersey number}
\item{Player}{ player's name} \item{GP}{ games played}
\item{GS}{ games started} \item{Min}{ minutes played}
\item{AvgMin}{ average minutes played per game}
\item{FG}{ field goals made} \item{FGA}{ field goals
attempted} \item{FGPct}{ field goal percentage}
\item{FG3}{ 3-point field goals made} \item{FG3A}{ 3-point
field goals attempted} \item{FG3Pct}{ 3-point field goal percentage}
\item{FT}{ free throws made} \item{FTA}{ free throws
attempted} \item{FTPct}{ free throw percentage}
\item{Off}{ offensive rebounds} \item{Def}{ defensive
rebounds} \item{Tot}{ total rebounds } \item{RBG}{ rebounds
per game} \item{PF}{ personal fouls} \item{FO}{ games fouled
out} \item{A}{ assists} \item{TO}{ turn overs}
\item{Blk}{ blocked shots} \item{Stl}{ steals}
\item{Pts}{ points scored} \item{PTSG}{ points per game} }}
\source{
MIAA sports archives (\url{http://www.miaa.org/})
}
\description{
Individual player statistics for the 2004-2005 Michigan Intercollegiate
Athletic Association basketball season.
}
\examples{

data(miaa05)
histogram(~FTPct, data=miaa05)

}
\keyword{datasets}

