% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pendulum}
\alias{Pendulum}
\title{Pendulum data}
\format{A data frame with 27 observations on the following 3 variables.
\itemize{ \item{length}{ length of the pendulum (in meters)}
\item{period}{ average time of period (in seconds) over several
swings of the pendulum} \item{delta.length}{ an estimate of the
accuracy of the length measurement} }}
\source{
Calvin College physics students under the direction of Professor
Steve Plath.
}
\description{
Period and pendulum length for a number of string and mass pendulums
constructed by physics students.  The same mass was used throughout, but the
length of the string was varied from 10cm to 16 m.
}
\examples{
data(Pendulum)
xyplot(period ~ length, data=Pendulum)
}
\keyword{datasets}

