% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxLik2.R
\name{plot.maxLik2}
\alias{plot.maxLik2}
\title{plot method for augment maxLik objects}
\usage{
\method{plot}{maxLik2}(x, y, ...)
}
\arguments{
\item{x}{an object of class \code{"maxLik2"}}

\item{y}{ignored}
}
\description{
See \code{\link{maxLik2}} and \code{\link{maxLik}} for how to create
the objects this method prints.
}

