% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RatPoison}
\alias{RatPoison}
\title{Rat poison -- unfinished documentation}
\format{
A data frame with 20 observations on the following 3 variables.
\itemize{ \item{consumption}{ a numeric vector}
\item{flavor}{ a factor with levels \code{bread}
\code{butter-vanilla} \code{plain} \code{roast beef}}
\item{location}{ a factor with levels \code{A} \code{B} \code{C}
\code{D} \code{E}} }
}
\description{
Data from an experiment to see whether flavor and location of rat poison
influence the consumption by rats.
}
\examples{

data(RatPoison)
gf_line(consumption ~ flavor, group = ~ location, color = ~ location, data = RatPoison) \%>\%
  gf_point()
}
\keyword{datasets}
