% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{rubberband}
\alias{rubberband}
\title{Rubber band launching -- unfinished documentation}
\format{A data frame with 16 observations on the following 2 variables.
\itemize{ \item{Stretch}{ amount rubber band was stretched before
launch} \item{Distance}{ distance rubber band traveled } }}
\description{
Results of an experiment comparing a rubber band travels to the amount it
was stretched prior to launch.
}
\examples{
data(rubberband)
xyplot(Distance ~ Stretch, data=rubberband, type=c("p","r"))
}
\keyword{datasets}

