% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RubberBand}
\alias{RubberBand}
\title{Rubber band launching -- unfinished documentation}
\format{
A data frame with 16 observations on the following 2 variables.
\itemize{ \item{Stretch}{ amount rubber band was stretched before
launch} \item{Distance}{ distance rubber band traveled } }
}
\description{
Results of an experiment comparing a rubber band travels to the amount it
was stretched prior to launch.
}
\examples{

data(RubberBand)
gf_point(distance ~ stretch, data = RubberBand) \%>\%
  gf_lm(interval = "confidence")

}
\keyword{datasets}
