% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Students}
\alias{Students}
\title{Standardized test scores and GPAs}
\format{A data frame with 1000 observations on the following 6 variables.
\itemize{ \item{ACT}{ ACT score} \item{SAT}{ SAT score}
\item{Grad}{ has the student graduated from college?}
\item{gradGPA}{ college GPA at graduation} \item{hsGPA}{ high
school GPA} \item{Cohort}{ year of graduation or expected graduation}
}}
\description{
Standardized test scores and GPAs for 1000 students.
}
\examples{

data(Students)
gf_point(ACT ~ SAT, data = Students)
gf_point(gradGPA ~ hsGPA, data = Students)

}
\keyword{datasets}
