% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{students}
\alias{students}
\title{Standardized test scores and GPAs}
\format{A data frame with 1000 observations on the following 6 variables.
\itemize{ \item{ACT}{ ACT score} \item{SAT}{ SAT score}
\item{Grad}{ has the student graduated from college?}
\item{GradGPA}{ college GPA at graduation} \item{HSGPA}{ high
school GPA} \item{Cohort}{ year of graduation or expected graduation}
}}
\description{
Standardized test scores and GPAs for 1000 students.
}
\examples{
data(students)
xyplot(ACT ~ SAT, data=students)
xyplot(GradGPA ~ HSGPA, data=students)
}
\keyword{datasets}

