% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undocumented.R
\name{undocumented}
\alias{funplot}
\alias{funvec}
\alias{grid.identify.points}
\alias{make.rm}
\alias{multirandom}
\alias{ntigram}
\alias{panel.lm}
\alias{panel.smooth}
\alias{panel.xyplotpoints}
\alias{panel.xyplotsmooth}
\alias{pdfplot}
\alias{pmfhistogram}
\alias{pmfplot}
\alias{print.stemplot}
\alias{probplot}
\alias{show.fumbles}
\alias{show.poisson}
\alias{stemplot}
\alias{stemplot.data.frame}
\alias{stemplot.numeric}
\alias{undocumented}
\title{Undocumented functions}
\description{
These objects are undocumented.
}
\details{
In many cases, these are depricated in favor of alternative
methods (often in the \code{mosaic} package) to obtain the same results but remain in 
the package to avoid breaking code used in Foundations and Applications of Statistics, the 
book associated with this package.
For example, \code{funplot}, should be replaced by \code{plotFun} and 
\code{pdfplot}, \code{pmfhistogram}, and \code{pmfplot} should be replaced by 
\code{plotDist}.

In other cases, the functions are of limited suitability for general use.
}
\author{
Randall Pruim
}

