% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undocumented.R
\name{undocumented}
\alias{funvec}
\alias{grid.identify.points}
\alias{undocumented}
\title{Undocumented functions}
\description{
These objects are undocumented.
}
\details{
Some are left-overs from a previous version of the book and package.
In other cases, the functions are of limited suitability for general use.
}
\author{
Randall Pruim
}

