% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaov.R
\name{vaov}
\alias{vaov}
\alias{vaov.formula}
\alias{vaov.formula}
\title{ANOVA vectors}
\usage{
vaov(x, ...)

\method{vaov}{formula}(x, data = parent.frame(), ...)
}
\arguments{
\item{x}{a formula.}

\item{\dots}{additional arguments.}

\item{data}{a data frame.}
}
\value{
A data frame with variables including \code{grandMean},
\code{groupMean}, \code{ObsVsGrand}, \code{STotal}, \code{ObsVsGroup},
\code{SError}, \code{GroupVsGrand}, and \code{STreatment}. The usual SS
terms can be computed from these by summing.
}
\description{
Compute vectors associated with 1-way ANOVA
}
\details{
This is primarily designed for demonstration purposes to show how 1-way
ANOVA models partition variance.  It may not work properly for more
complicated models.
}
\examples{

aov(pollution ~ location, data=airpollution)
vaov(pollution ~ location, data=airpollution)

}
\author{
Randall Pruim
}
\keyword{stats}
