% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WorkingWomen}
\alias{WorkingWomen}
\title{Women in the workforce}
\format{
A data frame with 19 observations on the following 3 variables.
\itemize{ \item{city}{ name of a U.S. city (coded as a factor with
19 levels)} \item{labor72}{ percent of women in labor force in 1972}
\item{labor68}{ percent of women in labor force in 1968} }
}
\source{
These data are from the United States Department of Labor Statistics
and are also available at DASL, the Data and Story Library
(\url{http://lib.stat.cmu.edu/DASL/}).
}
\description{
The labor force participation rate of women in each of 19 U.S. cities in
each of two years. # Reference: United States Department of Labor Statistics
# # Authorization: free use # # Description: # # Variable Names: # # 1.
City: City in the United States # 2. labor72: Labor Force Participation rate
of women in 1972 # 3. labor68: Labor Force Participation rate of women in
1968 # # The Data: #
}
\examples{

data(WorkingWomen)
gf_point(labor72 ~ labor68, data = WorkingWomen)

}
\keyword{datasets}
